feat(domain): add domain layer with snowflake IDs
- Introduce pkg/domain for clean architecture and business logic
- Add snowflake-based ID generator with InitIDGenerator and GenID
- Use int64 for NodeID in config to support distributed systems
- Model core entities: User, Product, Order, ChatMessage
- Implement repository interfaces as abstraction ports
- Use decimal.Decimal for monetary fields to prevent float errors
- Initialize domain layer in main during service startup

+++ pkg/config/config.go
@@ -49,7 +49,7 @@ type JwtConfig struct {
 }
 
 type SnowflakeConfig struct {
-	NodeID int `mapstructure:"node_id"`
+	NodeID int64 `mapstructure:"node_id"`
 }
 
 type LogConfig struct {
